refactor: modularize legal and app update features
