+++ lib/features/app_update/domain/models/app_update_info.dart
