+++ lib/features/events/presentation/map/events_map_page.dart
@@ -2,9 +2,9 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:crew_app/core/config/environment.dart';
-import 'package:crew_app/core/state/legal/disclaimer_providers.dart';
+import 'package:crew_app/features/legal/presentation/widgets/disclaimer_bottom_sheet.dart';
+import 'package:crew_app/features/legal/state/disclaimer_providers.dart';
 import 'package:crew_app/l10n/generated/app_localizations.dart';
-import 'package:crew_app/shared/legal/disclaimer_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_map/flutter_map.dart';
