+++ lib/features/legal/domain/models/disclaimer.dart
