+++ lib/features/settings/presentation/about/about_page.dart
@@ -1,4 +1,4 @@
-import 'package:crew_app/core/state/update/app_update_providers.dart';
+import 'package:crew_app/features/app_update/state/app_update_providers.dart';
 import 'package:crew_app/l10n/generated/app_localizations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
